refactor io module & add support for writing
This splits up the io module into a few separate files.  I'm not
convinced that I love the new organization, but it's a little better
than just throwing everything into lib.rs

Finally I added a quick and dirty `write_all` implmentation so that we
can write responses back to the client in our chat example so we have a
real echo server.